feat: create endpoints for frontend form
